Invalidate cache only on request by calling file handle builder's invalidateCache

+++ test/unit/org/apache/cassandra/io/util/WriteAndReadTest.java
@@ -25,13 +25,14 @@ import org.junit.Test;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import org.apache.cassandra.cache.ChunkCache;
+import org.apache.cassandra.config.Config;
 import org.apache.cassandra.config.DatabaseDescriptor;
+import org.apache.cassandra.db.compaction.OperationType;
 import org.apache.cassandra.io.compress.BufferType;
 import org.apache.cassandra.io.sstable.Component;
 import org.apache.cassandra.io.sstable.Descriptor;
 import org.apache.cassandra.io.sstable.SequenceBasedSSTableId;
-import org.apache.cassandra.utils.PageAware;
+import org.apache.cassandra.io.sstable.format.SSTableReaderBuilder;
 
 import static org.junit.Assert.assertEquals;
 
@@ -43,6 +44,7 @@ public class WriteAndReadTest
     public static void setupDD()
     {
         DatabaseDescriptor.daemonInitialization();
+        DatabaseDescriptor.setIndexAccessMode(Config.DiskAccessMode.standard);
     }
 
     @Test
@@ -60,9 +62,7 @@ public class WriteAndReadTest
 
         File parentDir = new File(System.getProperty("java.io.tmpdir"));
         Descriptor descriptor = new Descriptor(parentDir, "ks", "cf" + length, new SequenceBasedSSTableId(1));
-        try (FileHandle.Builder fhBuilder = new FileHandle.Builder(descriptor.fileFor(Component.PARTITION_INDEX))
-                                            .withChunkCache(ChunkCache.instance)
-                                            .bufferSize(PageAware.PAGE_SIZE))
+        try (FileHandle.Builder fhBuilder = SSTableReaderBuilder.primaryIndexWriteTimeBuilder(descriptor, Component.PARTITION_INDEX, OperationType.WRITE))
         {
             long root = length;
             long keyCount = root * length;
@@ -74,6 +74,7 @@ public class WriteAndReadTest
                             .trickleFsync(DatabaseDescriptor.getTrickleFsync())
                             .trickleFsyncByteInterval(DatabaseDescriptor.getTrickleFsyncIntervalInKb() * 1024)
                             .bufferType(BufferType.OFF_HEAP)
+                            .finishOnClose(true)
                             .build()))
 
             {
@@ -90,11 +91,12 @@ public class WriteAndReadTest
                 writer.sync();
             }
 
-            File filePath;
+            File filePath = null;
             // Now read it like PartitionIndex.load
             try (FileHandle fh = fhBuilder.complete();
                  FileDataInput rdr = fh.createReader(fh.dataLength() - 3 * 8))
             {
+                filePath = rdr.getFile();
                 long firstPosR = rdr.readLong();
                 long keyCountR = rdr.readLong();
                 long rootR = rdr.readLong();
@@ -103,10 +105,12 @@ public class WriteAndReadTest
                 assertEquals(keyCount, keyCountR);
                 assertEquals(rootR, root);
 
-                filePath = rdr.getFile();
             }
-
-            FileUtils.deleteWithConfirm(filePath);
+            finally
+            {
+                if (filePath != null)
+                    FileUtils.deleteWithConfirm(filePath);
+            }
         }
     }
 }
